requirements.txt
streamlit
pandas
scikit-learn
openai
google-generativeai
python-dotenv
